# optimise check of UTC / GMT during parsing