@omus 2020-09-04
Currently in master
```julia
julia> p1(name) = name in ("UTC", "GMT")
p1 (generic function with 1 method)

julia> @btime p1("ABC")
  15.051 ns (1 allocation: 32 bytes)
false

julia> @btime p1("UTC")
  9.519 ns (1 allocation: 32 bytes)
true

julia> @btime p1("GMT")
  12.639 ns (1 allocation: 32 bytes)
true
```

Current PR
```julia
julia> p2(name) = length(name) === 3 && @inbounds ((name[1] === 'U' && name[2]==='T' && name[3]==='C') || (name[1]==='G' && name[2]==='M' && name[3]==='T'))
p2 (generic function with 1 method)

julia> @btime p2("ABC")

  7.283 ns (0 allocations: 0 bytes)
false

julia> @btime p2("UTC")
  7.161 ns (0 allocations: 0 bytes)
true

julia> @btime p2("GMT")
  7.538 ns (0 allocations: 0 bytes)
true
```

Alternate
```julia
julia> p3(name) = name == "UTC" || name == "GMT"
p3 (generic function with 1 method)

julia> @btime p3("ABC")
  7.475 ns (0 allocations: 0 bytes)
false

julia> @btime p3("UTC")
  3.362 ns (0 allocations: 0 bytes)
true

julia> @btime p3("GMT")
  6.782 ns (0 allocations: 0 bytes)
true
```