@oxinabox 2020-09-04
I was not prepared for it to be `in` on a tuple of strings being the slow part.