@omus 2020-09-04
> Shall i reverse the commit that changed to `==`?

Let's try one more thing. What if you make `"UTC"` and `"GMT"` global constants?