@oxinabox 2020-09-04
>      Let's try one more thing. What if you make "UTC" and "GMT" global constants?

If that helps, we need to open a bug report on julia.