@oxinabox 2020-09-04
> Let's try one more thing. What if you make "UTC" and "GMT" global constants?

No it doesn't help.

```julia
julia> @btime parse(ZonedDateTime, "2017-11-14 11:03:53 +0100", dateformat"yyyy-mm-dd HH:MM:SS zzzzz");
  702.759 ns (13 allocations: 592 bytes)

julia> @btime parse(ZonedDateTime, "2016-04-11 08:00 UTC", dateformat"yyyy-mm-dd HH:MM ZZZ");
  672.657 ns (15 allocations: 704 bytes)

julia> @btime parse(ZonedDateTime, "2000+00", dateformat"yyyyz");
  504.984 ns (13 allocations: 496 bytes)

julia> @btime parse(ZonedDateTime, Test.GenericString("2018-01-01 00:00 UTC"), dateformat"yyyy-mm-dd HH:MM ZZZ");
  7.339 μs (127 allocations: 4.20 KiB)

julia> @btime tryparse(ZonedDateTime, "2013-03-20 11:00:00+04:00", dateformat"y-m-d H:M:SSz");
  878.184 ns (14 allocations: 640 bytes)

julia> @btime tryparse(ZonedDateTime, "2016-04-11 08:00 EST", dateformat"yyyy-mm-dd HH:MM zzz");
  90.995 ns (0 allocations: 0 bytes)

julia> @btime ZonedDateTime("2000-01-02T03:04:05.006+0700");
  832.155 ns (14 allocations: 688 bytes)

julia> @btime ZonedDateTime("2000-01-02T03:04:05.006Z");
  637.614 ns (9 allocations: 416 bytes)

julia> @btime ZonedDateTime("2018-11-01-0600", dateformat"yyyy-mm-ddzzzz");
  611.971 ns (13 allocations: 544 bytes)
```

---

I think the problem is the bencharks you used in 
https://github.com/JuliaTime/TimeZones.jl/pull/286#issuecomment-687254756
were using `Strings` 
where-as the real code
and the benchmark i did in https://github.com/JuliaTime/TimeZones.jl/issues/282
used `SubString`s.
and infact I did actually est in the original benchmarks `==`
it is under `is_UTC_or_GMT_or` (second result from top)