@omus 2020-09-04
Using `SubString`. 

Currently in master

```julia
julia> p1(name) = name in ("UTC", "GMT")
p1 (generic function with 1 method)

julia> @btime p1($(SubString("ABC")))
  49.681 ns (5 allocations: 224 bytes)
false

julia> @btime p1($(SubString("UTC")))
  36.300 ns (3 allocations: 128 bytes)
true

julia> @btime p1($(SubString("GMT")))
  57.718 ns (5 allocations: 224 bytes)
true
```

Current PR
```julia
julia> p2(name) = length(name) === 3 && @inbounds ((name[1] === 'U' && name[2]==='T' && name[3]==='C') || (name[1]==='G' && name[2]==='M' && name[3]==='T'))
p2 (generic function with 1 method)

julia> @btime p2($(SubString("ABC")))
  18.315 ns (0 allocations: 0 bytes)
false

julia> @btime p2($(SubString("UTC")))
  21.570 ns (0 allocations: 0 bytes)
true

julia> @btime p2($(SubString("GMT")))
  31.627 ns (0 allocations: 0 bytes)
true
```

Alternate
```julia
julia> p3(name) = name == "UTC" || name == "GMT"
p3 (generic function with 1 method)

julia> @btime p3($(SubString("ABC")))
  43.316 ns (4 allocations: 192 bytes)
false

julia> @btime p3($(SubString("UTC")))
  29.094 ns (2 allocations: 96 bytes)
true

julia> @btime p3($(SubString("GMT")))
  49.995 ns (4 allocations: 192 bytes)
true
```
There is probably a general performance fix we can do for all `SubString` equality then.