@omus 2020-09-04
The allocating seems to come from `==`. I made an alternate `==` which removes the allocations:
```julia
function alt(a::AbstractString, b::AbstractString)
    a === b && return true
    ncodeunits(a) != ncodeunits(b) && return false
    for (c::AbstractChar, d::AbstractChar) in zip(a, b)
        c != d && return false
    end
    return true
end
```


Current PR (Julia 1.5.1)

```julia
julia> p2(name) = length(name) === 3 && @inbounds ((name[1] === 'U' && name[2]==='T' && name[3]==='C') || (name[1]==='G' && name[2]==='M' && name[3]==='T'))
p2 (generic function with 1 method)

julia> @btime p2($(SubString("ABC")))

  18.557 ns (0 allocations: 0 bytes)
false

julia> @btime p2($(SubString("UTC")))
  21.320 ns (0 allocations: 0 bytes)
true

julia> @btime p2($(SubString("GMT")))
  25.836 ns (0 allocations: 0 bytes)
true
```

Alternate  (Julia 1.5.1)

```julia
julia> p4(name) = alt(name, "UTC") || alt(name, "GMT")
p4 (generic function with 1 method)

julia> @btime p4($(SubString("ABC")))

  13.796 ns (0 allocations: 0 bytes)
false

julia> @btime p4($(SubString("UTC")))
  13.799 ns (0 allocations: 0 bytes)
true

julia> @btime p4($(SubString("GMT")))
  19.570 ns (0 allocations: 0 bytes)
true
```

On Julia 1.6.0-DEV.842 (594ac899ff) it appears `==` doesn't allocate at all. I'll track down when it was fixed. For now I'd propose using the custom equality function which keeps the code easy to read.