@omus 2020-09-04
Looks like Julia 1.6.0-DEV.722 (https://github.com/JuliaLang/julia/pull/37192) was where the problem was fixed.