@omus 2020-09-08
Current PR (Julia 1.5.1)
```julia
julia> p3(name) = name == "UTC" || name == "GMT"
p3 (generic function with 1 method)

julia> @btime p3($(SubString("ABC")))
  8.193 ns (0 allocations: 0 bytes)
false

julia> @btime p3($(SubString("UTC")))
  6.663 ns (0 allocations: 0 bytes)
true

julia> @btime p3($(SubString("GMT")))
  9.507 ns (0 allocations: 0 bytes)
true
```